requirements.txt
streamlit>=1.28.0,<2
pandas>=2.0,<3
plotly>=5.18,<6
psycopg[binary]>=3.1,<4
python-dotenv>=1.0,<2
